+++ lib/haml_lint/linter/space_before_script.rb
@@ -48,37 +48,5 @@ module HamlLint
       text = node.value[:text].to_s
       text[0] != ' ' if text
     end
-
-    # Extracts all text for a tag node and normalizes it, including additional
-    # lines following commas or multiline bar indicators ('|')
-    def tag_with_inline_text(node)
-      # Next node is either the first child or the "next node" (node's sibling
-      # or next sibling of some ancestor)
-      next_node_line = [
-        [next_node(node), node.children.first].compact.map(&:line),
-        parser.lines.count + 1,
-      ].flatten.min
-
-      # Normalize each of the lines to ignore the multiline bar (|) and
-      # excess whitespace
-      parser.lines[(node.line - 1)...(next_node_line - 1)].map do |line|
-        line.strip.gsub(/\|\z/, '').rstrip
-      end.join(' ')
-    end
-
-    # Gets the next node following this node, ascending up the ancestor chain
-    # recursively if this node has no siblings.
-    def next_node(node)
-      return unless node
-      siblings = node.parent.children
-
-      next_sibling = siblings[siblings.index(node) + 1] if siblings.count > 1
-
-      if next_sibling
-        next_sibling
-      else
-        next_node(node.parent)
-      end
-    end
   end
 end